+++ lib/sec_id.rb
@@ -8,6 +8,7 @@ require 'sec_id/cusip'
 require 'sec_id/sedol'
 require 'sec_id/figi'
 require 'sec_id/lei'
+require 'sec_id/iban'
 require 'sec_id/cik'
 require 'sec_id/occ'